"fix" Webpack tree-shaking issue
