+++ packages/fluentui/react-bindings/src/compose/consts.ts
@@ -184,3 +184,33 @@ export type SlotProp<TProps> =
   | null
   | undefined
   | ObjectSlotProp<TProps>;
+
+//
+//
+//
+
+export const defaultComposeOptions: Required<ComposePreparedOptions> = {
+  className: process.env.NODE_ENV === 'production' ? '' : 'no-classname-🙉',
+  classes: [],
+  displayName: '',
+  displayNames: [],
+  mapPropsToStylesPropsChain: [],
+  render: () => null,
+  handledProps: [] as never[],
+  overrideStyles: false,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  slots: { __self: () => null },
+  slotProps: [],
+  state: props => props,
+  resolveSlotProps: () => ({}),
+  shorthandConfig: {},
+};
+
+/**
+ * A set of mapped props for intrinsic element types.
+ */
+export const defaultMappedProps: Record<string, string> = {
+  iframe: 'src',
+  img: 'src',
+  input: 'type',
+};
