--- packages/fluentui/react-bindings/src/compose/defaultMappedProps.ts
@@ -1,8 +1,0 @@
-/**
- * A set of mapped props for intrinsic element types.
- */
-export const defaultMappedProps: Record<string, string> = {
-  iframe: 'src',
-  img: 'src',
-  input: 'type',
-};
