+++ packages/fluentui/react-bindings/src/compose/index.ts
@@ -1,4 +1,4 @@
+export * from './consts';
 export { default as compose } from './compose';
 export { mergeProps } from './mergeProps';
 export { mergeSlotProp } from './mergeSlotProp';
-export * from './types';
